API "locations" endpoint

For detailed api ocumentation see https://transport.opendata.ch/docs.html#locations

locations(..., .protocol = "https")

Arguments

...
the locations endpoint requires the passing of either the query or the x and y parameters.

.protocol
the connection protocol; either http or https (default)

Value

the locations as a dataframe

Examples

# You need to specify either query or x and y
locations(query = "Bern")

#>         id                    name coordinate.x coordinate.y
#> 1  8507000                    Bern     46.94882     7.439123
#> 2  8576646           Bern, Bahnhof     46.94810     7.440213
#> 3  8579896    Bern, Hirschengraben     46.94606     7.437783
#> 4  8507110         Bern, Zytglogge     46.94784     7.447504
#> 5  8516161           Bern Wankdorf     46.96782     7.465469
#> 6  8588989        Bern, Bärenplatz     46.94812     7.442867
#> 7  8579897        Bern, Kocherpark     46.94588     7.434053
#> 8  8507061 Bern, Brunnadernstrasse     46.94118     7.463722
#> 9  8507060     Bern, Helvetiaplatz     46.94390     7.449210
#> 10 8507082       Bern  Europaplatz     46.94344     7.407470

locations(x = 46.94882, y = 7.439123)

#>         id                  name coordinate.x coordinate.y
#> 1     <NA>    Bern (Haltestelle)           NA           NA
#> 2  8507000                  Bern     46.94882     7.439123
#> 3  8507785    Bern, Hauptbahnhof     46.94899     7.438532
#> 4  8576646         Bern, Bahnhof     46.94810     7.440213
#> 5  8581351 Bern, Schanzenstrasse     46.94788     7.437074
#> 6  8579748     Bern, Obergericht     46.94997     7.436365
#> 7  8588989      Bern, Bärenplatz     46.94812     7.442867
#> 8  8579896  Bern, Hirschengraben     46.94606     7.437783
#> 9  8571357     Bern, Universität     46.95088     7.435590
#> 10 8588990        Bern, Bollwerk     46.95164     7.441370